ну тут довольно много поменялось

+++ Food.hpp
@@ -11,5 +11,5 @@ public:
     float get_amount() const;
     void set_amount(float value);
     void add(float value);
-    int take(int wanted_amount);
+    float take(float wanted_amount);
 };
